Create ilib-loctool-webos-common package and make to use it
