+++ packages/ilib-lint-webos/src/HtmlFormatter.js
@@ -1,7 +1,7 @@
 /*
  * HtmlFormatter.js - Formats result output
  *
- * Copyright © 2024 JEDLSoft
+ * Copyright (c) 2024 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
