+++ packages/ilib-lint-webos/src/index.js
@@ -1,7 +1,7 @@
 /*
  * index.js - main program of webos plugin
  *
- * Copyright © 2024 JEDLSoft
+ * Copyright (c) 2024 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
