+++ packages/ilib-lint-webos/test/webOSPlugin.test.js
@@ -1,7 +1,7 @@
 /*
  * webOSPlugin.test.js - test the webOS plugin
  *
- * Copyright © 2024 JEDLSoft
+ * Copyright (c) 2024 JEDLSoft
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
