+++ packages/ilib-loctool-webos-c/CFileType.js
@@ -21,8 +21,9 @@ var fs = require("fs");
 var path = require("path");
 var CFile = require("./CFile.js");
 var JsonResourceFileType = require("ilib-loctool-webos-json-resource");
-var Utils = require("loctool/lib/utils.js")
 var ResourceString = require("loctool/lib/ResourceString.js");
+var Utils = require("loctool/lib/utils.js")
+var pluginUtils = require("ilib-loctool-webos-common/utils.js");
 
 var CFileType = function(project) {
     this.type = "c";
@@ -97,17 +98,6 @@ CFileType.prototype._addResource = function(resFileType, translated, res, locale
     resource.setTargetLocale(locale);
     file = resFileType.getResourceFile(locale);
     file.addResource(resource);
-}
-
-CFileType.prototype._addNewResource = function(res, locale) {
-    var note = "No translation for " + res.reskey + " to " + locale;
-    var newres = res.clone();
-    newres.setTargetLocale(locale);
-    newres.setTarget(res.getSource());
-    newres.setState("new");
-    newres.setComment(note);
-    this.newres.add(newres);
-    this.logger.trace("No translation for " + res.reskey + " to " + locale);
 }
 
 /**
@@ -194,17 +184,17 @@ CFileType.prototype.write = function(translations, locales) {
                                             if (translated && (baseTranslation !== translated.getTarget())) {
                                                 this._addResource(resFileType, translated, res, locale);
                                             } else {
-                                                this._addNewResource(res, locale);
+                                                pluginUtils.addNewResource(this.newres, res, locale);
                                             }
                                         }.bind(this));
                                     } else if (translated && (baseTranslation !== translated.getTarget())){
                                         this._addResource(resFileType, translated, res, locale);
                                     } else {
-                                        this._addNewResource(res, locale);
+                                        pluginUtils.addNewResource(this.newres, res, locale);
                                     }
                                 }.bind(this));
                             } else {
-                                this._addNewResource(res, locale);
+                                pluginUtils.addNewResource(this.newres, res, locale);
                             }
                         }.bind(this));
                     } else if (!translated && customInheritLocale) {
@@ -212,7 +202,7 @@ CFileType.prototype.write = function(translations, locales) {
                             if (translated && (baseTranslation !== translated.getTarget())){
                                 this._addResource(resFileType, translated, res, locale);
                             } else {
-                                this._addNewResource(res, locale);
+                                pluginUtils.addNewResource(this.newres, res, locale);
                             }
                         }.bind(this));
                     } else if (!translated || ( this.API.utils.cleanString(res.getSource()) !== this.API.utils.cleanString(r.getSource()) &&
@@ -221,7 +211,7 @@ CFileType.prototype.write = function(translations, locales) {
                             this.logger.trace("extracted   source: " + this.API.utils.cleanString(res.getSource()));
                             this.logger.trace("translation source: " + this.API.utils.cleanString(r.getSource()));
                         }
-                        this._addNewResource(res, locale);
+                        pluginUtils.addNewResource(this.newres, res, locale);
                     } else {
                         if (res.reskey != r.reskey) {
                             // if reskeys don't match, we matched on cleaned string.
