+++ packages/ilib-loctool-webos-cpp/package.json
@@ -50,7 +50,8 @@
         "node": ">=10.0.0"
     },
     "dependencies": {
-        "ilib-loctool-webos-json-resource": "workspace:*"
+        "ilib-loctool-webos-json-resource": "workspace:*",
+        "ilib-loctool-webos-common": "workspace:*"
     },
     "peerDependencies": {
         "loctool": "2.28.2"
