+++ packages/ilib-loctool-webos-javascript/JavaScriptFileType.js
@@ -21,8 +21,9 @@ var fs = require("fs");
 var path = require("path");
 var JavaScriptFile = require("./JavaScriptFile.js");
 var JsonResourceFileType = require("ilib-loctool-webos-json-resource");
-var Utils = require("loctool/lib/utils.js")
 var ResourceString = require("loctool/lib/ResourceString.js");
+var Utils = require("loctool/lib/utils.js")
+var pluginUtils = require("ilib-loctool-webos-common/utils.js");
 
 var JavaScriptFileType = function(project) {
     this.type = "javascript";
@@ -119,16 +120,6 @@ JavaScriptFileType.prototype._addResource = function(resFileType, translated, re
     file.addResource(resource);
 }
 
-JavaScriptFileType.prototype._addNewResource = function(res, locale) {
-    var note = "No translation for " + res.reskey + " to " + locale;
-    var newres = res.clone();
-    newres.setTargetLocale(locale);
-    newres.setTarget(res.getSource());
-    newres.setState("new");
-    newres.setComment(note);
-    this.newres.add(newres);
-    this.logger.trace("No translation for " + res.reskey + " to " + locale);
-}
 
 /**
  * Write out the aggregated resources for this file type. In
@@ -216,18 +207,18 @@ JavaScriptFileType.prototype.write = function(translations, locales) {
                                             if (translated && (baseTranslation !== translated.getTarget())) {
                                                 this._addResource(resFileType, translated, res, locale);
                                             } else {
-                                                this._addNewResource(res, locale);
+                                                pluginUtils.addNewResource(this.newres, res, locale);
                                             }
                                         }.bind(this))
 
                                     } else if (translated && (baseTranslation !== translated.getTarget())){
                                         this._addResource(resFileType, translated, res, locale);
                                     } else {
-                                        this._addNewResource(res, locale);
+                                        pluginUtils.addNewResource(this.newres, res, locale);
                                     }
                                 }.bind(this));
                             } else {
-                                this._addNewResource(res, locale);
+                                pluginUtils.addNewResource(this.newres, res, locale);
                             }
                         }.bind(this));
                     } else if (!translated && customInheritLocale){
@@ -235,7 +226,7 @@ JavaScriptFileType.prototype.write = function(translations, locales) {
                             if (translated && (baseTranslation != translated.getTarget())) {
                                 this._addResource(resFileType, translated, res, locale);
                             } else {
-                                this._addNewResource(res, locale);
+                                pluginUtils.addNewResource(this.newres, res, locale);
                             }
                         }.bind(this));
                     } else if (!translated || ( this.API.utils.cleanString(res.getSource()) !== this.API.utils.cleanString(r.getSource()) &&
@@ -244,7 +235,7 @@ JavaScriptFileType.prototype.write = function(translations, locales) {
                             this.logger.trace("extracted   source: " + this.API.utils.cleanString(res.getSource()));
                             this.logger.trace("translation source: " + this.API.utils.cleanString(r.getSource()));
                         }
-                        this._addNewResource(res, locale);
+                        pluginUtils.addNewResource(this.newres, res, locale);
                     } else {
                         if (res.reskey != r.reskey) {
                             // if reskeys don't match, we matched on cleaned string.
