+++ packages/ilib-loctool-webos-qml/package.json
@@ -49,7 +49,8 @@
         "node": ">=10.0.0"
     },
     "dependencies": {
-        "ilib-loctool-webos-ts-resource": "workspace:*"
+        "ilib-loctool-webos-ts-resource": "workspace:*",
+        "ilib-loctool-webos-common": "workspace:*"
     },
     "peerDependencies": {
         "loctool": "2.28.2"
